feedy/requirements.txt
amqp==2.6.1
asgiref==3.2.10
attrs==19.3.0
billiard==3.6.3.0
celery==4.3.0
certifi==2020.6.20
chardet==3.0.4
Django==2.2.2
djangorestframework==3.11.1
djangorestframework-simplejwt==4.4.0
feedparser==5.2.1
idna==2.10
importlib-metadata==1.7.0
iniconfig==1.0.1
Inject==4.3.1
kombu==4.6.11
more-itertools==8.4.0
packaging==20.4
pluggy==0.13.1
py==1.9.0
PyJWT==1.7.1
pyparsing==2.4.7
pytest==6.0.1
pytz==2020.1
requests==2.24.0
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
urllib3==1.25.10
vine==1.3.0
zipp==3.1.0
eventlet==0.25.1
flower==0.9.3
greenlet==0.4.15
# psycopg2==2.7.6.1
psycopg2-binary==2.8.5
gunicorn==19.9.0
redis==3.5.3